callback refactor complete
