+++ include/string_util.h
@@ -1,18 +1,15 @@
+#ifndef _STRING_UTIL_H_
+#define _STRING_UTIL_H_
+
+#include <string>
+
 const std::string WHITESPACE = " \n\r\t\f\v";
 
-std::string ltrim(const std::string& s)
-{
-	size_t start = s.find_first_not_of(WHITESPACE);
-	return (start == std::string::npos) ? "" : s.substr(start);
-}
+std::string ltrim(const std::string& s);
 
-std::string rtrim(const std::string& s)
-{
-	size_t end = s.find_last_not_of(WHITESPACE);
-	return (end == std::string::npos) ? "" : s.substr(0, end + 1);
-}
+std::string rtrim(const std::string& s);
 
-std::string trim(const std::string& s)
-{
-	return rtrim(ltrim(s));
-}
+std::string trim(const std::string& s);
+
+
+#endif
